Move addmm utils to utils.py.
